+++ doki-theme-visualstudio/UITools.cs
@@ -10,14 +10,21 @@ namespace doki_theme_visualstudio {
       Func<DependencyObject, bool> predicate
     ) {
       while (true) {
-        var parent = childDependencyObject is Visual || childDependencyObject is Visual3D
-          ? VisualTreeHelper.GetParent(childDependencyObject)
-          : LogicalTreeHelper.GetParent(childDependencyObject);
+        var parent = GetParent(childDependencyObject);
         if (parent == null) return null;
 
         if (predicate(parent)) return parent;
         childDependencyObject = parent;
       }
+    }
+
+    public static DependencyObject? GetParent(
+      DependencyObject childDependencyObject
+    ) {
+      var parent = childDependencyObject is Visual || childDependencyObject is Visual3D
+        ? VisualTreeHelper.GetParent(childDependencyObject)
+        : LogicalTreeHelper.GetParent(childDependencyObject);
+      return parent;
     }
 
     public static void TraverseParentTree(